Beginnings of a refactor to split listening, sending, and modules
