+++ sync/class.jetpack-sync-full.php
@@ -9,7 +9,7 @@
  * - we fire an action called jetpack_full_sync_start so that WPCOM can erase the contents of the cached database
  * - for each object type, we obtain a full list of object IDs to sync via a single API call (hoping that since they're ints, they can all fit in RAM)
  * - we load the full objects for those IDs in chunks of Jetpack_Sync_Full::ARRAY_CHUNK_SIZE (to reduce the number of MySQL calls)
- * - we fire a trigger for the entire array which the Jetpack_Sync_Client then serializes and queues.
+ * - we fire a trigger for the entire array which the Jetpack_Sync_Sender then serializes and queues.
  */
 
 require_once 'class.jetpack-sync-wp-replicastore.php';
@@ -34,7 +34,7 @@ class Jetpack_Sync_Full {
 
 	// singleton functions
 	private static $instance;
-	private $client;
+	private $sender;
 
 	public static function getInstance() {
 		if ( null === self::$instance ) {
@@ -110,31 +110,31 @@ class Jetpack_Sync_Full {
 		return false;
 	}
 
-	private function get_client() {
-		if ( ! $this->client ) {
-			$this->client = Jetpack_Sync_Client::getInstance();
+	private function get_sender() {
+		if ( ! $this->sender ) {
+			$this->sender = Jetpack_Sync_Sender::getInstance();
 		}
 
-		return $this->client;
+		return $this->sender;
 	}
 
 	private function enqueue_all_constants() {
-		$total = $this->get_client()->full_sync_constants();
+		$total = $this->get_sender()->full_sync_constants();
 		$this->update_queue_progress( 'constants', $total );
 	}
 
 	private function enqueue_all_functions() {
-		$total = $this->get_client()->full_sync_callables();
+		$total = $this->get_sender()->full_sync_callables();
 		$this->update_queue_progress( 'functions', $total );
 	}
 
 	private function enqueue_all_options() {
-		$total = $this->get_client()->force_sync_options();
+		$total = $this->get_sender()->force_sync_options();
 		$this->update_queue_progress( 'options', $total );
 	}
 
 	private function enqueue_all_network_options() {
-		$total = $this->get_client()->force_sync_network_options();
+		$total = $this->get_sender()->force_sync_network_options();
 		$this->update_queue_progress( 'network_options', $total );
 	}
 
@@ -206,7 +206,7 @@ class Jetpack_Sync_Full {
 		$post_ids = $args[0];
 
 		$posts = array_map( array( 'WP_Post', 'get_instance' ), $post_ids );
-		$posts = array_map( array( $this->get_client(), 'filter_post_content_and_add_links' ), $posts );
+		$posts = array_map( array( $this->get_sender(), 'filter_post_content_and_add_links' ), $posts );
 
 		return array(
 			$posts,
@@ -258,7 +258,7 @@ class Jetpack_Sync_Full {
 
 	public function expand_options( $args ) {
 		if ( $args[0] ) {
-			return $this->get_client()->get_all_options();
+			return $this->get_sender()->get_all_options();
 		}
 
 		return $args;
@@ -266,14 +266,14 @@ class Jetpack_Sync_Full {
 
 	public function expand_constants( $args ) {
 		if ( $args[0] ) {
-			return $this->get_client()->get_all_constants();
+			return $this->get_sender()->get_all_constants();
 		}
 		return $args;
 	}
 
 	public function expand_callables( $args ) {
 		if ( $args[0] ) {
-			return $this->get_client()->get_all_callables();
+			return $this->get_sender()->get_all_callables();
 		}
 		return $args;
 	}
@@ -286,12 +286,12 @@ class Jetpack_Sync_Full {
 
 	public function expand_users( $args ) {
 		$user_ids = $args[0];
-		return array_map( array( $this->get_client(), 'sanitize_user_and_expand' ), get_users( array( 'include' => $user_ids ) ) );
+		return array_map( array( $this->get_sender(), 'sanitize_user_and_expand' ), get_users( array( 'include' => $user_ids ) ) );
 	}
 
 	public function expand_network_options( $args ) {
 		if ( $args[0] ) {
-			return $this->get_client()->get_all_network_options();
+			return $this->get_sender()->get_all_network_options();
 		}
 
 		return $args;
@@ -316,20 +316,20 @@ class Jetpack_Sync_Full {
 
 	// TODO:
 	private function enqueue_all_theme_info() {
-		$total = $this->get_client()->send_theme_info();
+		$total = $this->get_sender()->send_theme_info();
 		$this->update_queue_progress( 'themes', $total );
 	}
 
 	private function enqueue_all_updates() {
 
 		// check for updates
-		$total = $this->get_client()->full_sync_updates();
+		$total = $this->get_sender()->full_sync_updates();
 		$this->update_queue_progress( 'updates', $total );
 	}
 
 	public function expand_updates( $args ) {
 		if ( $args[0] ) {
-			return $this->get_client()->get_all_updates();
+			return $this->get_sender()->get_all_updates();
 		}
 
 		return $args;
