+++ tests/php/sync/test_class.jetpack-sync-constants.php
@@ -20,7 +20,7 @@ class WP_Test_Jetpack_New_Constants extends WP_Test_Jetpack_New_Sync_Base {
 		define( 'TEST_FOO', sprintf( "%.8f", microtime(true) ) );
 		define( 'TEST_BAR', sprintf( "%.8f", microtime(true) ) );
 
-		$this->client->do_sync();
+		$this->sender->do_sync();
 
 		$synced_foo_value = $this->server_replica_storage->get_constant( 'TEST_FOO' );
 		$synced_bar_value = $this->server_replica_storage->get_constant( 'TEST_BAR' );
@@ -31,7 +31,7 @@ class WP_Test_Jetpack_New_Constants extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_does_not_fire_if_constants_havent_changed() {
 		$this->client->set_defaults(); // use the default constants
-		$this->client->do_sync();
+		$this->sender->do_sync();
 
 		foreach( Jetpack_Sync_Defaults::$default_constants_whitelist as $constant ) {
 			try {
@@ -43,7 +43,7 @@ class WP_Test_Jetpack_New_Constants extends WP_Test_Jetpack_New_Sync_Base {
 		}
 
 		$this->server_replica_storage->reset();
-		$this->client->do_sync();
+		$this->sender->do_sync();
 
 		foreach( Jetpack_Sync_Defaults::$default_constants_whitelist as $constant ) {
 			$this->assertEquals( null, $this->server_replica_storage->get_constant( $constant ) );
@@ -52,16 +52,16 @@ class WP_Test_Jetpack_New_Constants extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_white_listed_constant_doesnt_get_synced_twice() {
 		$this->client->set_constants_whitelist( array( 'TEST_ABC' ) );
-		define( 'TEST_ABC', time() );
-		$this->client->do_sync();
+		define( 'TEST_ABC', microtime(true) );
+		$this->sender->do_sync();
 
 		$synced_value = $this->server_replica_storage->get_constant( 'TEST_ABC' );
-		$this->assertEquals( TEST_ABC, $synced_value );
+		$this->assertEquals( sprintf("%.2f", TEST_ABC), sprintf("%.2f", $synced_value ) );
 
 		$this->server_replica_storage->reset();
 		
-		delete_transient( Jetpack_Sync_Client::CONSTANTS_AWAIT_TRANSIENT_NAME );
-		$this->client->do_sync();
+		delete_transient( Jetpack_Sync_Module_Constants::CONSTANTS_AWAIT_TRANSIENT_NAME );
+		$this->sender->do_sync();
 
 		$this->assertEquals( null, $this->server_replica_storage->get_constant( 'TEST_ABC' ) );
 	}
