+++ tests/php/sync/test_class.jetpack-sync-modules.php
@@ -1,5 +1,4 @@
 <?php
-require_once dirname( __FILE__ ) . '/../../../sync/class.jetpack-sync-client.php';
 
 /**
  * Testing Activation and Deactivation of Modules
@@ -11,7 +10,7 @@ class WP_Test_Jetpack_Sync_Modules extends WP_Test_Jetpack_New_Sync_Base {
 		// Since the site need to eather be connected or in development mode.
 		// But we don't allow sync to happen in development mode.
 		do_action( 'jetpack_activate_module', 'stuff' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 
 		$events = $this->server_event_storage->get_all_events( 'jetpack_activate_module' );
 		$event = $events[0];
@@ -25,7 +24,7 @@ class WP_Test_Jetpack_Sync_Modules extends WP_Test_Jetpack_New_Sync_Base {
 		Jetpack_Options::update_option( 'active_modules', array( 'stuff' ) );
 		Jetpack::deactivate_module( 'stuff' );
 
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$events = $this->server_event_storage->get_all_events( 'jetpack_deactivate_module' );
 
 		$event = $events[0];
