+++ tests/php/sync/test_class.jetpack-sync-network-options.php
@@ -13,7 +13,7 @@ class WP_Test_Jetpack_New_Sync_Network_Options extends WP_Test_Jetpack_New_Sync_
 
 		$this->client->set_network_options_whitelist( array( 'test_network_option' ) );
 		add_site_option( 'test_network_option', 'foo' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 	}
 	public function tearDown() {
 		parent::tearDown();
@@ -32,7 +32,7 @@ class WP_Test_Jetpack_New_Sync_Network_Options extends WP_Test_Jetpack_New_Sync_
 			$this->markTestSkipped( 'Run it in multi site mode' );
 		}
 		update_site_option( 'test_network_option', 'bar' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$synced_option_value = $this->server_replica_storage->get_site_option( 'test_network_option' );
 		$this->assertEquals( 'bar', $synced_option_value );
 	}
@@ -42,7 +42,7 @@ class WP_Test_Jetpack_New_Sync_Network_Options extends WP_Test_Jetpack_New_Sync_
 			$this->markTestSkipped( 'Run it in multi site mode' );
 		}
 		delete_site_option( 'test_network_option' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$synced_option_value = $this->server_replica_storage->get_site_option( 'test_network_option' );
 		$this->assertEquals( false, $synced_option_value );
 	}
@@ -52,7 +52,7 @@ class WP_Test_Jetpack_New_Sync_Network_Options extends WP_Test_Jetpack_New_Sync_
 			$this->markTestSkipped( 'Run it in multi site mode' );
 		}
 		add_site_option( 'don_t_sync_test_network_option', 'foo' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$synced_option_value = $this->server_replica_storage->get_site_option( 'don_t_sync_test_network_option' );
 		$this->assertEquals( false, $synced_option_value );
 	}
