+++ tests/php/sync/test_class.jetpack-sync-sso.php
@@ -6,7 +6,7 @@
 class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 	function test_sync_sso_is_two_step_required_filter_true() {
 		add_filter( 'jetpack_sso_require_two_step', '__return_true' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_is_two_step_required' );
 		$this->assertTrue( $callableValue );
 		remove_filter( 'jetpack_sso_require_two_step', '__return_true' );
@@ -14,7 +14,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_should_hide_login_form_filter_true() {
 		add_filter( 'jetpack_remove_login_form', '__return_true' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_should_hide_login_form' );
 		$this->assertTrue( $callableValue );
 		remove_filter( 'jetpack_remove_login_form', '__return_true' );
@@ -22,7 +22,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_match_by_email_filter_true() {
 		add_filter( 'jetpack_sso_match_by_email', '__return_true' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_match_by_email' );
 		$this->assertTrue( $callableValue );
 		remove_filter( 'jetpack_sso_match_by_email', '__return_true' );
@@ -30,7 +30,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_new_user_override_filter_true() {
 		add_filter( 'jetpack_sso_new_user_override', '__return_true' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_new_user_override' );
 		$this->assertTrue( $callableValue );
 		remove_filter( 'jetpack_sso_new_user_override', '__return_true' );
@@ -38,7 +38,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_sso_bypass_default_login_form_filter_true() {
 		add_filter( 'jetpack_sso_bypass_login_forward_wpcom', '__return_true' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_bypass_default_login_form' );
 		$this->assertTrue( $callableValue );
 		remove_filter( 'jetpack_sso_bypass_login_forward_wpcom', '__return_true' );
@@ -46,7 +46,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_is_two_step_required_filter_false() {
 		add_filter( 'jetpack_sso_require_two_step', '__return_false' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_is_two_step_required' );
 		$this->assertFalse( $callableValue );
 		remove_filter( 'jetpack_sso_require_two_step', '__return_false' );
@@ -54,7 +54,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_should_hide_login_form_filter_false() {
 		add_filter( 'jetpack_remove_login_form', '__return_false' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_should_hide_login_form' );
 		$this->assertFalse( $callableValue );
 		remove_filter( 'jetpack_remove_login_form', '__return_false' );
@@ -62,7 +62,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_match_by_email_filter_false() {
 		add_filter( 'jetpack_sso_match_by_email', '__return_false' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_match_by_email' );
 		$this->assertFalse( $callableValue );
 		remove_filter( 'jetpack_sso_match_by_email', '__return_false' );
@@ -70,7 +70,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_new_user_override_filter_false() {
 		add_filter( 'jetpack_sso_new_user_override', '__return_false' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_new_user_override' );
 		$this->assertFalse( $callableValue );
 		remove_filter( 'jetpack_sso_new_user_override', '__return_false' );
@@ -78,7 +78,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_sso_bypass_default_login_form_filter_false() {
 		add_filter( 'jetpack_sso_bypass_login_forward_wpcom', '__return_false' );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_bypass_default_login_form' );
 		$this->assertFalse( $callableValue );
 		remove_filter( 'jetpack_sso_bypass_login_forward_wpcom', '__return_false' );
@@ -86,7 +86,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_is_two_step_required_option_true() {
 		update_option( 'jetpack_sso_require_two_step', true );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_is_two_step_required' );
 		$this->assertTrue( $callableValue );
 		delete_option( 'jetpack_sso_require_two_step' );
@@ -94,7 +94,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_should_hide_login_form_option_true() {
 		update_option( 'jetpack_sso_remove_login_form', true );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_should_hide_login_form' );
 		$this->assertTrue( $callableValue );
 		delete_option( 'jetpack_sso_remove_login_form' );
@@ -102,7 +102,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_is_two_step_required_option_false() {
 		update_option( 'jetpack_sso_require_two_step', false );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_is_two_step_required' );
 		$this->assertFalse( $callableValue );
 		delete_option( 'jetpack_sso_require_two_step' );
@@ -110,7 +110,7 @@ class WP_Test_Jetpack_Sync_SSO extends WP_Test_Jetpack_New_Sync_Base {
 
 	function test_sync_sso_should_hide_login_form_option_false() {
 		update_option( 'jetpack_sso_remove_login_form', false );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$callableValue = $this->server_replica_storage->get_callable( 'sso_should_hide_login_form' );
 		$this->assertFalse( $callableValue );
 		delete_option( 'jetpack_sso_remove_login_form' );
