+++ tests/php/sync/test_class.jetpack-sync-updates.php
@@ -10,7 +10,7 @@ class WP_Test_Jetpack_New_Sync_Updates extends WP_Test_Jetpack_New_Sync_Base {
 		parent::setUp();
 		$this->client->reset_data();
 		wp_set_current_user( 1 );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 	}
 
 	public function test_update_plugins_is_synced() {
@@ -19,7 +19,7 @@ class WP_Test_Jetpack_New_Sync_Updates extends WP_Test_Jetpack_New_Sync_Base {
 		}
 		
 		wp_update_plugins();
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$updates = $this->server_replica_storage->get_updates( 'plugins' );
 
 		$this->assertFalse( isset( $updates->no_update ) );
@@ -34,7 +34,7 @@ class WP_Test_Jetpack_New_Sync_Updates extends WP_Test_Jetpack_New_Sync_Base {
 		}
 
 		wp_update_themes();
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$updates = $this->server_replica_storage->get_updates( 'themes' );
 		$this->assertTrue( is_int( $updates->last_checked ) );
 	}
@@ -45,7 +45,7 @@ class WP_Test_Jetpack_New_Sync_Updates extends WP_Test_Jetpack_New_Sync_Base {
 		}
 
 		_maybe_update_core();
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$updates = $this->server_replica_storage->get_updates( 'core' );
 		$this->assertTrue( is_int( $updates->last_checked ) );
 	}
@@ -57,7 +57,7 @@ class WP_Test_Jetpack_New_Sync_Updates extends WP_Test_Jetpack_New_Sync_Base {
 		// Lets pretend that we updated the wp_version to bar.
 		$wp_version = 'bar';
 		do_action( 'upgrader_process_complete', null, array( 'action' => 'update', 'type' => 'core' ) );
-		$this->client->do_sync();
+		$this->sender->do_sync();
 		$this->assertEquals( $wp_version, $this->server_replica_storage->get_callable( 'wp_version' ) );
 	}
